vcs: Write ignore files before output
Previously, when creating temporary output directories tytanic would create the
directories, write the output and then write the ignore file. This is fine in
theory but doesn't play well with a VCS like jujutsu if a watchman trigger is
used because it will track the changes before adding the ignore file.

Closes #87.