filter: Disallow commas an parentheses raw patterns
Most often a pattenr is used with a fairly limited set of characters, trying
to use these directly inside arguments or expression groups would require extra
whitespace or quoting for very simple cases.

Raw patterns no longer parse `,`, `(` and `)`, which allows them to be used
without the extra whitespace at the cost of quoting for regexes with capture
groups.