[wip] core: Add a world builder
Splitting the different parts of the `World` trait into differnet parts
makes it easier to swap out individual details like it is required for
template tests and their special file lookup.

However, this also came at the cost of making the old access check for
tmeplates hard to achieve without piercing the abstraction boundary of
those traits.

TODO: This change still has some unwraps due to how template proiders
are loaded, in particular a typed version of a template project would
be good to avoid this.