core: Add a world builder
Splitting the different parts of the `World` trait into differnet parts
makes it easier to swap out individual details like it is required for
template tests and their special file lookup.

However, this also came at the cost of making the old access check for
templates hard to achieve without piercing the abstraction boundary of
those traits. This is mostly due to the fact that we can't emit any
non-fatal diagnostics from the `World` implementation.