core: Ensure correct transitive file lookup
Because the `PackageSpec` was previously removed for re-routes to the
template package root itself the typst internal world lookup created
`FileId`s for this root which had no spec either. This obviously creates
paths which the provider can no longer distinguish from normal template
paths.

Incidentally, this fix also allows for more overrides to be placed in
the provider directly which could be used for patching dependencies in
tests. However, this also comes with the downside of not allowing the
re-routing with a virtual prvoider, but this was not previously used.

Fixes #229