fix: Assume default value for project config defaults if undefined
Rather than running into a deserialization error. Previously every field
of the (nested) `ProjectConfig` had a serde `default` instruction,
except for (ironically) the `ProjectDefaults` sub-structure. This lead
to the weird situation where users of this package had to add an
(essentially empty) `tool.tytanic.default` section to their `typst.toml`
to fix config parsing when overwriting, for example, the
`tool.tytanic.tests` key.

Now the `ProjectDefaults` are deserialized to their `Default` impl if
not present and there is a test to catch future regressions of the same
kind.